Convert offsets from int to double

+++ Source/Extensions/Blazorise.SignaturePad/Models/SignaturePadBeginStrokeEventArgs.cs
@@ -10,7 +10,7 @@ public class SignaturePadBeginStrokeEventArgs
     /// </summary>
     /// <param name="offsetX">The X-axis offset of the stroke.</param>
     /// <param name="offsetY">The Y-axis offset of the stroke.</param>
-    public SignaturePadBeginStrokeEventArgs( int offsetX, int offsetY )
+    public SignaturePadBeginStrokeEventArgs( double offsetX, double offsetY )
     {
         OffsetX = offsetX;
         OffsetY = offsetY;
@@ -19,10 +19,10 @@ public class SignaturePadBeginStrokeEventArgs
     /// <summary>
     /// Gets the X-axis offset of the stroke.
     /// </summary>
-    public int OffsetX { get; }
+    public double OffsetX { get; }
 
     /// <summary>
     /// Gets the Y-axis offset of the stroke.
     /// </summary>
-    public int OffsetY { get; }
+    public double OffsetY { get; }
 }
